Fix what I *think* is a long-standing bug.
Previously: The class has a dev_class field which isn't used. It also has a
setDeviceClass method, which is used internally, but is misleadingly documented
to imply this sets the device class on startup. The actual thing that defines
device class is the _CLASS_C define, and that unconditionally makes the device
into a class C device.

I've changed things so that setDeviceClass now just sets dev_class. The original
setDeviceClass has been renamed to be a private function, setRunningDeviceClass,
and this is now used internally. The checks for _CLASS_C are now replaced with
checks against the dev_class value.

I guess this makes the code a bit bigger on class A devices - maybe I should
put a #define back in so that you can completely remove all class-dependent code.